// ev-backoffice/client/src/app/core/models/benefit-categories.enum.ts
export enum BenefitCategories {
  IR1 = 'IR1',
  IR2 = 'IR2',
  IR3 = 'IR3',
  IR4 = 'IR4',
  IR5 = 'IR5',
  F1_A = 'F1_A',
  F2_A = 'F2_A',
  F3_A = 'F3_A',
  F4_A = 'F4_A',
  F1_B = 'F1_B',
  F2_B = 'F2_B',
  F3_B = 'F3_B',
  F4_B = 'F4_B',
  K1K3 = 'K1K3',
  K2K4 = 'K2K4',
  K1K3AOS = 'K1K3AOS',
  K2K4AOS = 'K2K4AOS',
  AOS = 'AOS',
  NATURALIZATION = 'NATURALIZATION',
  LPRSPOUSE = 'LPRSPOUSE',
  LPRCHILD = 'LPRCHILD',
  SIX01A = 'SIX01A',
  SIX01 = 'SIX01',
  EAD = 'EAD',
  DISABILITY = 'DISABILITY',
  REMOVECOND = 'REMOVECOND',
}