// ev-backoffice/client/src/app/core/models/benefit-groups.enum.ts
export enum BenefitGroups {
  IMMEDIATE_RELATIVE_VISA = 'IMMEDIATE_RELATIVE_VISA',
  VISA_FIANCE_SPOUSE = 'VISA_FIANCE_SPOUSE',
  PERMANENT_RESIDENCE = 'PERMANENT_RESIDENCE',
  REMOVE_CONDITIONS = 'REMOVE_CONDITIONS',
  FAMILY_PREFERENCE_VISA = 'FAMILY_PREFERENCE_VISA',
  NATURALIZATION = 'NATURALIZATION',
  MISCELLANEOUS = 'MISCELLANEOUS',
}